move device_meta to own file, add MetaGroup
