+++ qcodes/instrument/group_parameter.py
@@ -155,6 +155,8 @@ class Group:
             values (as directly obtained from the output of the get command;
             note that parsers within the parameters will take care of
             individual parsing of their values).
+        single_instrument: A flag to indicate that all parameters belong to a
+        single instrument, which in turn does additional checks. Defaults to True.
     """
     def __init__(self,
                  parameters: List[GroupParameter],
@@ -162,16 +164,18 @@ class Group:
                  get_cmd: Optional[str] = None,
                  get_parser: Union[Callable[[str],
                                             Dict[str, Any]], None] = None,
-                 separator: str = ','
+                 separator: str = ',',
+                 single_instrument: bool = True
                  ) -> None:
         self._parameters = OrderedDict((p.name, p) for p in parameters)
 
         for p in parameters:
             p._group = self
 
-        if len({p.root_instrument for p in parameters}) > 1:
-            raise ValueError(
-                "All parameters should belong to the same instrument")
+        if single_instrument:
+            if len({p.root_instrument for p in parameters}) > 1:
+                raise ValueError(
+                    "All parameters should belong to the same instrument")
 
         self._instrument = parameters[0].root_instrument
 
@@ -182,7 +186,11 @@ class Group:
             self.get_parser = get_parser
         else:
             self.get_parser = self._separator_parser(separator)
+        
+        if single_instrument:
+            self._check_initial_values(parameters)
 
+    def _check_initial_values(self, parameters):
         have_initial_values = [p._initial_value is not None
                                for p in parameters]
         if any(have_initial_values):
