+++ qcodes/instrument_drivers/meta/__init__.py
@@ -5,6 +5,7 @@ from qcodes.instrument_drivers.meta.meta_instrument import (
 
 from qcodes.instrument_drivers.meta.chip_meta import (
     ChipMeta,
-    DeviceMeta,
-    DeviceMetaParameter
+    DeviceMeta
 )
+
+from qcodes.instrument_drivers.meta.device_meta import DeviceMetaParameter
